requirements.txt
python-telegram-bot==20.3
fastapi==0.111.0
uvicorn==0.20.0
requests==2.32.3
